requirements.txt
requests
beautifulsoup4
pydantic
PyYAML
python-dateutil
playwright